backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
pydantic>=2.9,<3
pydantic-settings==2.2.1
google-generativeai==0.7.2
httpx==0.27.0
starlette==0.37.2
